Moving file based session property manager classes to a separate package
