+++ presto-session-property-managers/src/main/java/io/prestosql/plugin/session/file/FileSessionPropertyManagerFactory.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.prestosql.plugin.session;
+package io.prestosql.plugin.session.file;
 
 import com.google.inject.Injector;
 import io.airlift.bootstrap.Bootstrap;
