+++ presto-session-property-managers/src/test/java/io/prestosql/plugin/session/file/TestFileSessionPropertyManager.java
@@ -12,9 +12,11 @@
  * limitations under the License.
  */
 
-package io.prestosql.plugin.session;
+package io.prestosql.plugin.session.file;
 
 import io.airlift.testing.TempFile;
+import io.prestosql.plugin.session.AbstractTestSessionPropertyManager;
+import io.prestosql.plugin.session.SessionMatchSpec;
 import io.prestosql.spi.session.SessionPropertyConfigurationManager;
 
 import java.io.IOException;
@@ -23,7 +25,7 @@ import java.nio.file.Path;
 import java.util.Arrays;
 import java.util.Map;
 
-import static io.prestosql.plugin.session.FileSessionPropertyManager.CODEC;
+import static io.prestosql.plugin.session.file.FileSessionPropertyManager.CODEC;
 import static org.testng.Assert.assertEquals;
 
 public class TestFileSessionPropertyManager
