+++ presto-session-property-managers/src/test/java/io/prestosql/plugin/session/file/TestFileSessionPropertyManagerConfig.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.prestosql.plugin.session;
+package io.prestosql.plugin.session.file;
 
 import com.google.common.collect.ImmutableMap;
 import org.testng.annotations.Test;
